moved: Domain related classes to domain module and network related classes to network module
