--- app/src/main/java/dev/enesky/doodle/core/network/di/RepositoryModule.kt
@@ -1,22 +1,0 @@
-package dev.enesky.doodle.core.network.di
-
-import dev.enesky.doodle.core.network.api.service.JikanService
-import dev.enesky.doodle.core.network.paging.PopularAnimesPagingSource
-import dev.enesky.doodle.core.network.repository.JikanDataSource
-import dev.enesky.doodle.core.network.repository.JikanRepository
-import org.koin.dsl.module
-
-/**
- * Created by Enes Kamil YILMAZ on 28/10/2023
- */
-
-val repositoryModule = module {
-
-    single<JikanRepository> {
-        JikanRepository(get<JikanDataSource>())
-    }
-
-    single<PopularAnimesPagingSource> {
-        PopularAnimesPagingSource(get<JikanService>())
-    }
-}
