--- app/src/main/java/dev/enesky/doodle/core/network/util/Resource.kt
@@ -1,21 +1,0 @@
-package dev.enesky.doodle.core.network.util
-
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onStart
-
-sealed interface Resource<out T> {
-    data class Success<T>(val data: T) : Resource<T>
-    data class Error(val exception: Throwable? = null) : Resource<Nothing>
-    object Loading : Resource<Nothing>
-}
-
-fun <T> Flow<T>.asResource(): Flow<Resource<T>> {
-    return this
-        .map<T, Resource<T>> {
-            Resource.Success(it)
-        }
-        .onStart { emit(Resource.Loading) }
-        .catch { emit(Resource.Error(it)) }
-}
