--- app/src/main/java/dev/enesky/doodle/core/util/Constants.kt
@@ -1,9 +1,0 @@
-package dev.enesky.doodle.core.util
-
-/**
- * Created by Enes Kamil YILMAZ on 28/10/2023
- */
-object Constants {
-
-    const val ITEMS_PER_PAGE = 25
-}
