+++ core/domain/src/main/java/dev/enesky/core/domain/di/UseCaseModule.kt
@@ -1,0 +1,26 @@
+package dev.enesky.core.domain.di
+
+import dev.enesky.core.network.repository.JikanRepository
+import dev.enesky.core.domain.usecase.AnimeCharactersUseCase
+import dev.enesky.core.domain.usecase.AnimeUseCase
+import dev.enesky.core.domain.usecase.PopularAnimesUseCase
+import org.koin.dsl.module
+
+/**
+ * Created by Enes Kamil YILMAZ on 28/10/2023
+ */
+
+val useCaseModule = module {
+
+    single<PopularAnimesUseCase> {
+        PopularAnimesUseCase(get<JikanRepository>()::getPopularAnimes)
+    }
+
+    single<AnimeUseCase> {
+        AnimeUseCase(get<JikanRepository>())
+    }
+
+    single<AnimeCharactersUseCase> {
+        AnimeCharactersUseCase { get<JikanRepository>().getCharactersByAnimeId(it) }
+    }
+}
