+++ core/network/src/main/java/dev/enesky/core/network/repository/JikanDataSource.kt
@@ -1,0 +1,17 @@
+package dev.enesky.core.network.repository
+
+import androidx.paging.PagingData
+import dev.enesky.core.network.model.Anime
+import dev.enesky.core.network.model.Character
+import kotlinx.coroutines.flow.Flow
+
+/**
+ * Created by Enes Kamil YILMAZ on 28/10/2023
+ */
+
+interface JikanDataSource {
+
+    fun getPopularAnimes(): Flow<PagingData<Anime>>
+    suspend fun getAnimeById(animeId: Int): Result<Anime>
+    suspend fun getCharactersByAnimeId(animeId: Int): Result<List<Character>>
+}
