+++ core/network/src/main/java/dev/enesky/core/network/repository/JikanRepository.kt
@@ -1,0 +1,28 @@
+package dev.enesky.core.network.repository
+
+import androidx.paging.PagingData
+import dev.enesky.core.network.model.Anime
+import dev.enesky.core.network.model.Character
+import dev.enesky.core.network.repository.JikanDataSource
+import kotlinx.coroutines.flow.Flow
+
+/**
+ * Created by Enes Kamil YILMAZ on 28/10/2023
+ */
+
+class JikanRepository(
+    private val jikanDataSource: JikanDataSource
+) {
+
+    fun getPopularAnimes(): Flow<PagingData<Anime>> {
+        return jikanDataSource.getPopularAnimes()
+    }
+
+    suspend fun getAnimeById(animeId: Int): Result<Anime> {
+        return jikanDataSource.getAnimeById(animeId)
+    }
+
+    suspend fun getCharactersByAnimeId(animeId: Int): Result<List<Character>> {
+        return jikanDataSource.getCharactersByAnimeId(animeId)
+    }
+}
