+++ core/network/src/main/java/dev/enesky/core/network/util/Constants.kt
@@ -1,0 +1,10 @@
+package dev.enesky.core.network.util
+
+/**
+ * Created by Enes Kamil YILMAZ on 28/10/2023
+ */
+object Constants {
+
+    const val ITEMS_PER_PAGE = 25
+
+}
